Use separate listening sockets for IPv6 and IPv4
For hairtunes.c and shairport.pl, use separate sockets for listening
on IPv4 and IPv6. This fixes a problem with IPv4 connectivity on OpenBSD.

Previously, a single INET6 socket was used to listen on both IPv4 and IPv6.
This works properly whenever the IPV6_V6ONLY option is set to
false, and IPV6_V6ONLY defaults to false on most operating systems.

However, on OpenBSD this option is permanently set 'true' (for security
reasons), so the IPv6 sockets were only listening on IPv6 addresses
(hairtunes/shairport were not listening on IPv4 at all).

Thus, separate IPv4 sockets are needed to listen on IPv4 addresses.